requirements.txt
opencv-python
numpy
open3d
scikit-learn
scipy
fastapi
uvicorn
Pillow
pyvista
requests
